refactor dynamics, add paths as arguments
