+++ workspace/src/localization/localization_py/localization_py/4DOF_dynamics.yml
@@ -6,9 +6,3 @@ i_wheel: .001
 gamma: 0.33333333
 tau_0: 0.3
 omega_0: 30
-
-Q1: 0.1
-Q3: 3
-Q4: 0.1
-R1: 0.0
-R3: 0.3
